Moved TabletOrder, it's not client-only
