+++ src/main/java/mcvmcomputers/ClientMod.java
@@ -27,10 +27,10 @@ import mcvmcomputers.client.entities.render.KeyboardRender;
 import mcvmcomputers.client.entities.render.MouseRender;
 import mcvmcomputers.client.entities.render.PCRender;
 import mcvmcomputers.client.tablet.TabletOS;
-import mcvmcomputers.client.tablet.TabletOrder;
 import mcvmcomputers.entities.EntityItemPreview;
 import mcvmcomputers.entities.EntityList;
 import mcvmcomputers.networking.PacketList;
+import mcvmcomputers.utils.TabletOrder;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendereregistry.v1.EntityRendererRegistry;
 import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
