+++ src/main/java/mcvmcomputers/client/tablet/TabletOS.java
@@ -23,7 +23,6 @@ import io.netty.buffer.Unpooled;
 import mcvmcomputers.ClientMod;
 import mcvmcomputers.MainMod;
 import mcvmcomputers.client.entities.model.OrderingTabletModel;
-import mcvmcomputers.client.tablet.TabletOrder.OrderStatus;
 import mcvmcomputers.entities.EntityDeliveryChest;
 import mcvmcomputers.item.ItemList;
 import mcvmcomputers.item.OrderableItem;
@@ -31,6 +30,7 @@ import mcvmcomputers.networking.PacketList;
 import mcvmcomputers.sound.SoundList;
 import mcvmcomputers.sound.TabletSoundInstance;
 import mcvmcomputers.utils.MVCUtils;
+import mcvmcomputers.utils.TabletOrder.OrderStatus;
 import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.sound.PositionedSoundInstance;
