+++ src/main/java/mcvmcomputers/networking/PacketList.java
@@ -12,10 +12,10 @@ import java.util.zip.Inflater;
 import io.netty.buffer.Unpooled;
 import mcvmcomputers.ClientMod;
 import mcvmcomputers.MainMod;
-import mcvmcomputers.client.tablet.TabletOrder;
 import mcvmcomputers.entities.EntityPC;
 import mcvmcomputers.item.ItemHarddrive;
 import mcvmcomputers.item.OrderableItem;
+import mcvmcomputers.utils.TabletOrder;
 import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
 import net.fabricmc.fabric.api.network.ServerSidePacketRegistry;
 import net.fabricmc.fabric.api.server.PlayerStream;
